test 000error

+++ src/mbot_linux_serial.cpp
@@ -79,7 +79,7 @@ void write_pose_and_orientation(double positionX, double positionY, double posit
     // 通过串口下发数据
     boost::asio::write(sp, boost::asio::buffer(buf));
     for(int i=0;i<62;i++)
-    {cout<<hex<<buf[i]<<endl; }
+    {cout<<hex<<(int)buf[i]<<endl; }
     
 }
 /********************************************************
